Changes mostly in typecast to prevent warnings from compiler.
git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/trunk@201 c60273ff-4a12-0410-8d58-82ceea6d4170

+++ sphinx2/src/examples/raw2cep.c
@@ -324,7 +324,7 @@ main (int32 argc, char **argv)
 	float *mfcp;
 
 	memset(&param, 0, sizeof(param));
-	param.SAMPLING_RATE = sps;
+	param.SAMPLING_RATE = (float)sps;
 
 	if ((fe = fe_init (&param)) == NULL)
 	    E_FATAL("fe_init(%d) failed\n", sps);
